# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeGate",
    "description": "Generative AI CodeGen security gateway",
    "version": "0.1.7"
  },
  "paths": {
...
    "/api/v1/provider-endpoints": {
      "get": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "List Provider Endpoints",
        "description": "List all provider endpoints.",
        "operationId": "v1_list_provider_endpoints",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderEndpoint"
                  },
                  "title": "Response V1 List Provider Endpoints"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "Add Provider Endpoint",
        "description": "Add a provider endpoint.",
        "operationId": "v1_add_provider_endpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderEndpoint"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEndpoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider-endpoints/{provider_id}": {
      "get": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "Get Provider Endpoint",
        "description": "Get a provider endpoint by ID.",
        "operationId": "v1_get_provider_endpoint",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEndpoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "Update Provider Endpoint",
        "description": "Update a provider endpoint by ID.",
        "operationId": "v1_update_provider_endpoint",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderEndpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEndpoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "Delete Provider Endpoint",
        "description": "Delete a provider endpoint by id.",
        "operationId": "v1_delete_provider_endpoint",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider-endpoints/{provider_name}/models": {
      "get": {
        "tags": [
          "CodeGate API",
          "Providers"
        ],
        "summary": "List Models By Provider",
        "description": "List models by provider.",
        "operationId": "v1_list_models_by_provider",
        "parameters": [
          {
            "name": "provider_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelByProvider"
                  },
                  "title": "Response V1 List Models By Provider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "List Workspaces",
        "description": "List all workspaces.",
        "operationId": "v1_list_workspaces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWorkspacesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Create Workspace",
        "description": "Create a new workspace.",
        "operationId": "v1_create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrRenameWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/active": {
      "get": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "List Active Workspaces",
        "description": "List all active workspaces.\n\nIn it's current form, this function will only return one workspace. That is,\nthe globally active workspace.",
        "operationId": "v1_list_active_workspaces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListActiveWorkspacesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Activate Workspace",
        "description": "Activate a workspace by name.",
        "operationId": "v1_activate_workspace",
        "parameters": [
          {
            "name": "status_code",
            "in": "query",
            "required": false,
            "schema": {
              "default": 204,
              "title": "Status Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_name}": {
      "delete": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Delete Workspace",
        "description": "Delete a workspace by name.",
        "operationId": "v1_delete_workspace",
        "parameters": [
          {
            "name": "workspace_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/archive/{workspace_name}/recover": {
      "post": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Recover Workspace",
        "description": "Recover an archived workspace by name.",
        "operationId": "v1_recover_workspace",
        "parameters": [
          {
            "name": "workspace_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/archive/{workspace_name}": {
      "delete": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Hard Delete Workspace",
        "description": "Hard delete an archived workspace by name.",
        "operationId": "v1_hard_delete_workspace",
        "parameters": [
          {
            "name": "workspace_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_name}/alerts": {
      "get": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Get Workspace Alerts",
        "description": "Get alerts for a workspace.",
        "operationId": "v1_get_workspace_alerts",
        "parameters": [
          {
            "name": "workspace_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/AlertConversation"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response V1 Get Workspace Alerts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelByProvider": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider"
        ],
        "title": "ModelByProvider",
        "description": "Represents a model supported by a provider.\n\nNote that these are auto-discovered by the provider."
      },
...
      "ProviderEndpoint": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "provider_type": {
            "$ref": "#/components/schemas/ProviderType"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "auth_type": {
            "$ref": "#/components/schemas/ProviderAuthType"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "provider_type",
          "endpoint",
          "auth_type"
        ],
        "title": "ProviderEndpoint",
        "description": "Represents a provider's endpoint configuration. This\nallows us to persist the configuration for each provider,\nso we can use this for muxing messages."
      },
...
    }
  }
}